brod (Brod v3.15.4)

Types

batch_input()

Specs

t:brod.batch_input/0

bootstrap()

Specs

t:brod.bootstrap/0

call_ref()

Specs

t:brod.call_ref/0

cg()

Specs

t:brod.cg/0

cg_protocol_type()

Specs

t:brod.cg_protocol_type/0

client()

Specs

t:brod.client/0

client_config()

Specs

t:brod.client_config/0

client_id()

Specs

t:brod.client_id/0

compression()

Specs

t:brod.compression/0

conn_config()

Specs

t:brod.conn_config/0

connection()

Specs

t:brod.connection/0

consumer_config()

Specs

t:brod.consumer_config/0

consumer_option()

Specs

t:brod.consumer_option/0

consumer_options()

Specs

t:brod.consumer_options/0

endpoint()

Specs

t:brod.endpoint/0

error_code()

Specs

t:brod.error_code/0

fetch_opts()

Specs

t:brod.fetch_opts/0

fold_acc()

Specs

t:brod.fold_acc/0

fold_fun(acc)

Specs

t:brod.fold_fun/1

fold always returns when reaches the high watermark offset. fold also returns when any of the limits is hit.

fold_limits()

Specs

t:brod.fold_limits/0

fold_result()

Specs

t:brod.fold_result/0

fold_stop_reason()

Specs

t:brod.fold_stop_reason/0

OffsetToContinue: begin offset for the next fold call

group_config()

Specs

t:brod.group_config/0

group_generation_id()

Specs

t:brod.group_generation_id/0

group_id()

Specs

t:brod.group_id/0

group_member()

Specs

t:brod.group_member/0

group_member_id()

Specs

t:brod.group_member_id/0

hostname()

Specs

t:brod.hostname/0

key()

Specs

t:brod.key/0

message()

Specs

t:brod.message/0

message_set()

Specs

t:brod.message_set/0

msg_input()

Specs

t:brod.msg_input/0

msg_ts()

Specs

t:brod.msg_ts/0

offset()

Specs

t:brod.offset/0

offset_time()

Specs

t:brod.offset_time/0

partition()

Specs

t:brod.partition/0

partition_assignment()

Specs

t:brod.partition_assignment/0

partition_fun()

Specs

t:brod.partition_fun/0

partitioner()

Specs

t:brod.partitioner/0

portnum()

Specs

t:brod.portnum/0

produce_ack_cb()

Specs

t:brod.produce_ack_cb/0

produce_reply()

Specs

t:brod.produce_reply/0

produce_result()

Specs

t:brod.produce_result/0

producer_config()

Specs

t:brod.producer_config/0

received_assignments()

Specs

t:brod.received_assignments/0

topic()

Specs

t:brod.topic/0

topic_config()

Specs

t:brod.topic_config/0

topic_partition()

Specs

t:brod.topic_partition/0

value()

Specs

t:brod.value/0

Functions

connect_group_coordinator(BootstrapEndpoints, ConnCfg, GroupId)

Specs

brod.connect_group_coordinator/3

Connect to consumer group coordinator broker.
Done in steps:

Connect to any of the given bootstrap ednpoints

Send group_coordinator_request to resolve group coordinator endpoint

Connect to the resolved endpoint and return the connection pid

connect_leader(Hosts, Topic, Partition, ConnConfig)

Specs

brod.connect_leader/4

Connect partition leader.

consume_ack(ConsumerPid, Offset)

Specs

brod.consume_ack/2

consume_ack(Client, Topic, Partition, Offset)

Specs

brod.consume_ack/4

create_topics(Hosts, TopicConfigs, RequestConfigs)

Specs

brod.create_topics/3

Equivalent to create_topics(Hosts, TopicsConfigs, RequestConfigs, []).

create_topics(Hosts, TopicConfigs, RequestConfigs, Options)

Specs

brod.create_topics/4

Create topic(s) in kafka Return the message body of create_topics, response. See kpro_schema.erl for struct details

delete_topics(Hosts, Topics, Timeout)

Specs

brod.delete_topics/3

Equivalent to delete_topics(Hosts, Topics, Timeout, []).

delete_topics(Hosts, Topics, Timeout, Options)

Specs

brod.delete_topics/4

Delete topic(s) from kafka Return the message body of delete_topics, response. See kpro_schema.erl for struct details

describe_groups(CoordinatorEndpoint, ConnCfg, IDs)

Specs

brod.describe_groups/3

Describe consumer groups. The given consumer group IDs should be all managed by the coordinator-broker running at the given endpoint. Otherwise error codes will be returned in the result structs. Return describe_groups response body field named groups. See kpro_schema.erl for struct details

fetch(ConnOrBootstrap, Topic, Partition, Offset)

Specs

brod.fetch/4

Fetch a single message set from the given topic-partition. The first arg can either be an already established connection to leader, or {Endpoints, ConnConfig} so to establish a new connection before fetch.

fetch(ConnOrBootstrap, Topic, Partition, Offset, Opts)

Specs

brod.fetch/5

Fetch a single message set from the given topic-partition. The first arg can either be an already established connection to leader, or {Endpoints, ConnConfig} so to establish a new connection before fetch.

fetch(Hosts, Topic, Partition, Offset, MaxWaitTime, MinBytes, MaxBytes)

This function is deprecated. fetch(Hosts, Topic, Partition, Offset, Wait, MinBytes, MaxBytes, []).

Specs

brod.fetch/7

fetch(Hosts, Topic, Partition, Offset, MaxWaitTime, MinBytes, MaxBytes, ConnConfig)

This function is deprecated. Fetch a single message set from the given topic-partition..

Specs

brod.fetch/8

fetch_committed_offsets(Client, GroupId)

Specs

brod.fetch_committed_offsets/2

Same as fetch_committed_offsets/3, but works with a started brod_client

fetch_committed_offsets(BootstrapEndpoints, ConnCfg, GroupId)

Specs

brod.fetch_committed_offsets/3

Fetch committed offsets for ALL topics in the given consumer group. Return the responses field of the offset_fetch response. See kpro_schema.erl for struct details.

fold(Bootstrap, Topic, Partition, Offset, Opts, Acc, Fun, Limits)

Specs

brod.fold/8

Fold through messages in a partition. Works like lists:foldl/2 but with below stop conditions: * Always return after reach high watermark offset * Return after the given message count limit is reached * Return after the given kafka offset is reached. * Return if the FoldFun returns an {error, Reason} tuple. NOTE: Exceptions from evaluating FoldFun are not caught.

get_consumer(Client, Topic, Partition)

Specs

brod.get_consumer/3

get_metadata(Hosts)

Specs

brod.get_metadata/1

Fetch broker metadata Return the message body of metadata response. See kpro_schema.erl for details

get_metadata(Hosts, Topics)

Specs

brod.get_metadata/2

Fetch broker/topic metadata Return the message body of metadata response. See kpro_schema.erl for struct details

get_metadata(Hosts, Topics, Options)

Specs

brod.get_metadata/3

Fetch broker/topic metadata Return the message body of metadata response. See kpro_schema.erl for struct details

get_partitions_count(Client, Topic)

Specs

brod.get_partitions_count/2

Get number of partitions for a given topic. The higher level producers may need the partition numbers to find the partition producer pid --- if the number of partitions is not statically configured for them. It is up to the callers how they want to distribute their data (e.g. random, roundrobin or consistent-hashing) to the partitions.

get_producer(Client, Topic, Partition)

Specs

brod.get_producer/3

Equivalent to brod_client:get_producer / 3.

list_all_groups(Endpoints, ConnCfg)

Specs

brod.list_all_groups/2

List ALL consumer groups in the given kafka cluster. NOTE: Exception if failed to connect any of the coordinator brokers.

list_groups(CoordinatorEndpoint, ConnCfg)

Specs

brod.list_groups/2

List consumer groups in the given group coordinator broker.

produce(Pid, Value)

Specs

brod.produce/2

Equivalent to produce(Pid, 0, <<>>, Value).

produce(ProducerPid, Key, Value)

Specs

brod.produce/3

Produce one message if Value is a binary or an iolist. Otherwise send a batch, if Value is a (nested) key-value list, or a list of maps. In this case Key is discarded (only the keys in the key-value list are sent to Kafka). The pid should be a partition producer pid, NOT client pid. The return value is a call reference of type call_ref(), so the caller can use it to expect (match) a #brod_produce_reply{result = brod_produce_req_acked} message after the produce request has been acked by Kafka.

produce(Client, Topic, Partition, Key, Value)

Specs

brod.produce/5

Produce one message if Value is a binary or an iolist. Otherwise send a batch if Value is a (nested) key-value list, or a list of maps. In this case Key is used only for partitioning, or discarded if the 3rd argument is a partition number instead of a partitioner callback. This function first looks up the producer pid, then calls produce/3 to do the real work. The return value is a call reference of type call_ref(), so the caller can used it to expect (match) a #brod_produce_reply{result = brod_produce_req_acked} message after the produce request has been acked by Kafka.

produce_cb(ProducerPid, Key, Value, AckCb)

Specs

brod.produce_cb/4

Same as produce/3, only the ack is not delivered as a message, instead, the callback is evaluated by producer worker when ack is received from kafka.

produce_cb(Client, Topic, Part, Key, Value, AckCb)

Specs

brod.produce_cb/6

Same as produce/5 only the ack is not delivered as a message, instead, the callback is evaluated by producer worker when ack is received from kafka. Return the partition to caller as {ok, Partition} for caller to correlate the callback when the 3rd arg is not a partition number.

produce_no_ack(ProducerPid, Key, Value)

Specs

brod.produce_no_ack/3

Send the message to partition worker without any ack. NOTE: This call has no back-pressure to the caller, excessive usage may cause BEAM to run out of memory.

produce_no_ack(Client, Topic, Part, Key, Value)

Specs

brod.produce_no_ack/5

Find the partition worker and send message without any ack. NOTE: This call has no back-pressure to the caller, excessive usage may cause BEAM to run out of memory.

produce_sync(Pid, Value)

Specs

brod.produce_sync/2

Equivalent to produce_sync(Pid, 0, <<>>, Value).
Same as produce/5 only the ack is not d

produce_sync(Pid, Key, Value)

Specs

brod.produce_sync/3

Sync version of produce/3
This function will not return until the response is received from Kafka. But when producer is started with required_acks set to 0, this function will return once the messages are buffered in the producer process.

produce_sync(Client, Topic, Partition, Key, Value)

Specs

brod.produce_sync/5

Sync version of produce/5 This function will not return until a response is received from kafka, however if producer is started with required_acks set to 0, this function will return once the messages are buffered in the producer process.

produce_sync_offset(Client, Topic, Partition, Key, Value)

Specs

brod.produce_sync_offset/5

Version of produce_sync/5 that returns the offset assigned by Kafka If producer is started with required_acks set to 0, the offset will be ?BROD_PRODUCE_UNKNOWN_OFFSET.

resolve_offset(Hosts, Topic, Partition)

Specs

brod.resolve_offset/3

Equivalent to resolve_offset(Hosts, Topic, Partition, latest, 1).

resolve_offset(Hosts, Topic, Partition, Time)

Specs

brod.resolve_offset/4

Resolve semantic offset or timestamp to real offset.

resolve_offset(Hosts, Topic, Partition, Time, ConnCfg)

Specs

brod.resolve_offset/5

Resolve semantic offset or timestamp to real offset.

resolve_offset(Hosts, Topic, Partition, Time, ConnCfg, Opts)

Specs

brod.resolve_offset/6

Resolve semantic offset or timestamp to real offset.

start()

Specs

brod.start/0

Start brod application.

start(StartType, StartArgs)

Application behaviour callback

start_client(BootstrapEndpoints)

Specs

brod.start_client/1

Equivalent to start_client(BootstrapEndpoints, brod_default_client).

start_client(BootstrapEndpoints, ClientId)

Specs

brod.start_client/2

Equivalent to start_client(BootstrapEndpoints, ClientId, []).

start_client(BootstrapEndpoints, ClientId, Config)

Specs

brod.start_client/3

Start a client.

BootstrapEndpoints: Kafka cluster endpoints, can be any of the brokers in the cluster, which does not necessarily have to be the leader of any partition, e.g. a load-balanced entrypoint to the remote Kafka cluster.

ClientId: Atom to identify the client process.
Config is a proplist, possible values:

restart_delay_seconds (optional, default=10)
How long to wait between attempts to restart brod_client process when it crashes.

get_metadata_timeout_seconds (optional, default=5)
Return {error, timeout} from brod_client:get_xxx calls if responses for APIs such as metadata, find_coordinator are not received in time.

reconnect_cool_down_seconds (optional, default=1)
Delay this configured number of seconds before retrying to estabilish a new connection to the kafka partition leader.

allow_topic_auto_creation (optional, default=true)
By default, brod respects what is configured in the broker about topic auto-creation. i.e. whether auto.create.topics.enable is set in the broker configuration. However if allow_topic_auto_creation is set to false in client config, brod will avoid sending metadata requests that may cause an auto-creation of the topic regardless of what broker config is.

auto_start_producers (optional, default=false)
If true, brod client will spawn a producer automatically when user is trying to call produce but did not call brod:start_producer explicitly. Can be useful for applications which don't know beforehand which topics they will be working with.

default_producer_config (optional, default=[])
Producer configuration to use when auto_start_producers is true. See brod_producer:start_link/4 for details about producer config

Connection options can be added to the same proplist. See kpro_connection.erl in kafka_protocol for the details:

ssl (optional, default=false)
true | false | ssl:ssl_option()true is translated to [] as ssl:ssl_option() i.e. all default.

sasl (optional, default=undefined)
Credentials for SASL/Plain authentication. {mechanism(), Filename} or {mechanism(), UserName, Password} where mechanism can be atoms: plain (for "PLAIN"), scram_sha_256 (for "SCRAM-SHA-256") or scram_sha_512 (for SCRAM-SHA-512). Filename should be a file consisting two lines, first line is the username and the second line is the password. Both Username and Password should be string() | binary()

connect_timeout (optional, default=5000)
Timeout when trying to connect to an endpoint.

request_timeout (optional, default=240000, constraint: >= 1000)
Timeout when waiting for a response, connection restart when timed out.

query_api_versions (optional, default=true)
Must be set to false to work with kafka versions prior to 0.10, When set to true, at connection start, brod will send a query request to get the broker supported API version ranges. When set to 'false', brod will alway use the lowest supported API version when sending requests to kafka. Supported API version ranges can be found in: brod_kafka_apis:supported_versions/1

extra_sock_opts (optional, default=[])
Extra socket options to tune socket performance. e.g. [{sndbuf, 1 bsl 20}]. More info

start_consumer(Client, TopicName, ConsumerConfig)

Specs

brod.start_consumer/3

Dynamically start a topic consumer.
See also: for details about consumer config..

start_link_client(BootstrapEndpoints)

Specs

brod.start_link_client/1

Equivalent to start_link_client(BootstrapEndpoints, brod_default_client).

start_link_client(BootstrapEndpoints, ClientId)

Specs

brod.start_link_client/2

Equivalent to start_link_client(BootstrapEndpoints, ClientId, []).

start_link_client(BootstrapEndpoints, ClientId, Config)

Specs

brod.start_link_client/3

start_link_group_subscriber(Client, GroupId, Topics, GroupConfig, ConsumerConfig, CbModule, CbInitArg)

Specs

brod.start_link_group_subscriber/7

See also: brod_group_subscriber:start_link/7.

start_link_group_subscriber(Client, GroupId, Topics, GroupConfig, ConsumerConfig, MessageType, CbModule, CbInitArg)

Specs

brod.start_link_group_subscriber/8

See also: brod_group_subscriber:start_link/8.

start_link_group_subscriber_v2(Config)

Specs

brod.start_link_group_subscriber_v2/1

Start group_subscriber_v2

start_link_topic_subscriber(Config)

Specs

brod.start_link_topic_subscriber/1

See also: brod_topic_subscriber:start_link/1.

start_link_topic_subscriber(Client, Topic, ConsumerConfig, CbModule, CbInitArg)

This function is deprecated. Please use start_link_topic_subscriber/1 instead.

Specs

brod.start_link_topic_subscriber/5

Equivalent to start_link_topic_subscriber(Client, Topic, all, ConsumerConfig, CbModule, CbInitArg).

start_link_topic_subscriber(Client, Topic, Partitions, ConsumerConfig, CbModule, CbInitArg)

This function is deprecated. Please use start_link_topic_subscriber/1 instead.

Specs

brod.start_link_topic_subscriber/6

Equivalent to start_link_topic_subscriber(Client, Topic, Partitions, ConsumerConfig, message, CbModule, CbInitArg).

start_link_topic_subscriber(Client, Topic, Partitions, ConsumerConfig, MessageType, CbModule, CbInitArg)

This function is deprecated. Please use start_link_topic_subscriber/1 instead.

Specs

brod.start_link_topic_subscriber/7

See also: brod_topic_subscriber:start_link/7.

start_producer(Client, TopicName, ProducerConfig)

Specs

brod.start_producer/3

Dynamically start a per-topic producer.
See also: brod_producer:start_link/4.

stop()

Specs

brod.stop/0

Stop brod application.

stop(State)

Application behaviour callback

stop_client(Client)

Specs

brod.stop_client/1

Stop a client.

subscribe(ConsumerPid, SubscriberPid, Options)

Specs

brod.subscribe/3

subscribe(Client, SubscriberPid, Topic, Partition, Options)

Specs

brod.subscribe/5

Subscribe to a data stream from the given topic-partition.

If {error, Reason} is returned, the caller should perhaps retry later.

{ok, ConsumerPid} is returned on success. The caller may want to monitor the consumer pid and re-subscribe should the ConsumerPid crash.

Upon successful subscription the subscriber process should expect messages of pattern: {ConsumerPid, #kafka_message_set{}} and {ConsumerPid, #kafka_fetch_error{}}.

-include_lib("brod/include/brod.hrl") to access the records.
In case #kafka_fetch_error{} is received the subscriber should re-subscribe itself to resume the data stream.

sync_produce_request(CallRef)

Specs

brod.sync_produce_request/1

Block wait for sent produced request to be acked by kafka.

sync_produce_request(CallRef, Timeout)

Specs

brod.sync_produce_request/2

sync_produce_request_offset(CallRef)

Specs

brod.sync_produce_request_offset/1

As sync_produce_request_offset/1, but also returning assigned offset See produce_sync_offset/5.

sync_produce_request_offset(CallRef, Timeout)

Specs

brod.sync_produce_request_offset/2

unsubscribe(ConsumerPid)

Specs

brod.unsubscribe/1

Unsubscribe the current subscriber. Assuming the subscriber is self().

unsubscribe(ConsumerPid, SubscriberPid)

Specs

brod.unsubscribe/2

Unsubscribe the current subscriber.

unsubscribe(Client, Topic, Partition)

Specs

brod.unsubscribe/3

Unsubscribe the current subscriber. Assuming the subscriber is self().

unsubscribe(Client, Topic, Partition, SubscriberPid)

Specs

brod.unsubscribe/4

Unsubscribe the current subscriber.