brod_cg_commits (Brod v3.15.4)

This is a utility module to help force commit offsets to kafka.

Types

group_id()

Specs

t:brod_cg_commits.group_id/0

group_input()

Specs

t:brod_cg_commits.group_input/0

member_id()

Specs

t:brod_cg_commits.member_id/0

offset()

Specs

t:brod_cg_commits.offset/0

offsets()

Specs

t:brod_cg_commits.offsets/0

partition()

Specs

t:brod_cg_commits.partition/0

prop_key()

Specs

t:brod_cg_commits.prop_key/0

prop_val()

Specs

t:brod_cg_commits.prop_val/0

retention()

Specs

t:brod_cg_commits.retention/0

-1 to use whatever configured in kafka

topic()

Specs

t:brod_cg_commits.topic/0

Functions

assign_all_to_self(Members, TopicPartitions)

Specs

brod_cg_commits.assign_all_to_self/2

assign_partitions(Pid, Members, TopicPartitionList)

Specs

brod_cg_commits.assign_partitions/3

This function is called only when partition_assignment_strategy is set for callback_implemented in group config.

assignments_received(Pid, MemberId, GenerationId, TopicAssignments)

Specs

brod_cg_commits.assignments_received/4

Called by group coordinator when there is new assignemnt received.

assignments_revoked(Pid)

Specs

brod_cg_commits.assignments_revoked/1

Called by group coordinator before re-joinning the consumer group.

code_change(OldVsn, State, Extra)

Callback implementation for :gen_server.code_change/3.

get_committed_offsets(Pid, TopicPartitions)

Specs

brod_cg_commits.get_committed_offsets/2

Called by group coordinator when initializing the assignments for subscriber. NOTE: this function is called only when it is DISABLED to commit offsets to kafka. i.e. offset_commit_policy is set to consumer_managed

handle_call(Call, From, State0)

Callback implementation for :gen_server.handle_call/3.

handle_cast(Cast, State)

Callback implementation for :gen_server.handle_cast/2.

handle_info(Info, State)

Callback implementation for :gen_server.handle_info/2.

init(_)

Callback implementation for :gen_server.init/1.

log(State, Level, Fmt, Args)

maybe_reply_sync(State)

run(ClientId, GroupInput)

Force commit offsets.

set_done(State)

start_link(Client, GroupInput)

Specs

brod_cg_commits.start_link/2

Start (link) a group member. The member will try to join the consumer group and get assignments for the given topic-partitions, then commit given offsets to kafka. In case not all given partitions are assigned to it, it will terminate with an exit exception

stop(Pid)

Specs

brod_cg_commits.stop/1

Stop the process.

sync(Pid)

Specs

brod_cg_commits.sync/1

Make a call to the resetter process, the call will be blocked until offsets are committed.

terminate(Reason, State)

Callback implementation for :gen_server.terminate/2.