brod_client (Brod v3.15.4)

Types

client()

Specs

t:brod_client.client/0

client_id()

Specs

t:brod_client.client_id/0

config()

Specs

t:brod_client.config/0

conn_state()

Specs

t:brod_client.conn_state/0

connection()

Specs

t:brod_client.connection/0

dead_conn()

Specs

t:brod_client.dead_conn/0

endpoint()

Specs

t:brod_client.endpoint/0

get_consumer_error()

Specs

t:brod_client.get_consumer_error/0

get_producer_error()

Specs

t:brod_client.get_producer_error/0

get_worker_error()

Specs

t:brod_client.get_worker_error/0

group_id()

Specs

t:brod_client.group_id/0

partition()

Specs

t:brod_client.partition/0

partition_worker_key()

Specs

t:brod_client.partition_worker_key/0

state()

Specs

t:brod_client.state/0

topic()

Specs

t:brod_client.topic/0

Functions

config(Key, Config, Default)

conn_config(State)

conn_config(Rest, ConnCfgKeys, Acc)

connect(State, Endpoint)

Specs

brod_client.connect/2

deregister_consumer(Client, Topic, Partition)

Specs

brod_client.deregister_consumer/3

De-register the consumer for a partition. The partition consumer entry is deleted from the ETS table to allow cleanup of purposefully stopped consumers and allow later restart.

deregister_producer(Client, Topic, Partition)

Specs

brod_client.deregister_producer/3

De-register the producer for a partition. The partition producer entry is deleted from the ETS table to allow cleanup of purposefully stopped producers and allow later restart.

do_connect(Endpoint, State)

do_get_group_coordinator(State0, GroupId)

Specs

brod_client.do_get_group_coordinator/2

do_get_leader_connection(State0, Topic, Partition)

do_get_metadata(Topic, State)

Specs

brod_client.do_get_metadata/2

do_get_partitions_count(TopicMetadata)

Specs

brod_client.do_get_partitions_count/1

do_start_consumer(TopicName, ConsumerConfig, State)

do_start_producer(TopicName, ProducerConfig, State)

ensure_binary(ClientId)

ensure_metadata_connection(State)

ensure_partition_workers(TopicName, State, F)

find_conn(Endpoint, Conns)

Specs

brod_client.find_conn/2

find_consumer(Client, Topic, Partition)

Specs

brod_client.find_consumer/3

find_partition_worker(Client, _)

Specs

brod_client.find_partition_worker/2

find_producer(Client, Topic, Partition)

Specs

brod_client.find_producer/3

get_connection(Client, Host, Port)

Specs

brod_client.get_connection/3

Get connection to a kafka broker. Return already established connection towards the broker, otherwise a new one is established and cached in client state. If the old connection was dead less than a configurable N seconds ago, {error, LastReason} is returned.

get_consumer(Client, Topic, Partition)

Specs

brod_client.get_consumer/3

Get consumer of the given topic-parition.

get_group_coordinator(Client, GroupId)

Specs

brod_client.get_group_coordinator/2

Get broker endpoint and connection config for connecting a group coordinator.

get_leader_connection(Client, Topic, Partition)

Specs

brod_client.get_leader_connection/3

Get the connection to kafka broker which is a leader for given Topic-Partition. Return already established connection towards the leader broker, Otherwise a new one is established and cached in client state. If the old connection was dead less than a configurable N seconds ago, {error, LastReason} is returned.

get_metadata(Client, Topic)

Specs

brod_client.get_metadata/2

Get topic metadata, if topic is undefined, will fetch ALL metadata.

get_metadata_connection(State)

get_metadata_safe(Topic0, State)

Specs

brod_client.get_metadata_safe/2

get_partition_worker(ClientPid, Key)

Specs

brod_client.get_partition_worker/2

get_partitions_count(Client, Topic)

Specs

brod_client.get_partitions_count/2

Get number of partitions for a given topic.

get_partitions_count(Client, Ets, Topic)

Specs

brod_client.get_partitions_count/3

get_producer(Client, Topic, Partition)

Specs

brod_client.get_producer/3

Get producer of the given topic-partition. The producer is started if auto_start_producers is enabled in client config.

handle_connection_down(State, Pid, Reason)

Specs

brod_client.handle_connection_down/3

is_cooled_down(Ts, State)

kf(FieldName, Struct)

Specs

brod_client.kf/2

lookup_partition_worker(Client, Ets, Key)

Specs

brod_client.lookup_partition_worker/3

lookup_partitions_count_cache(Ets, Topic)

Specs

brod_client.lookup_partitions_count_cache/2

mark_dead(Reason)

maybe_connect(State, Endpoint)

Specs

brod_client.maybe_connect/2

maybe_connect(State, Endpoint, Reason)

Specs

brod_client.maybe_connect/3

maybe_start_producer(Client, Topic, Partition, Error)

Specs

brod_client.maybe_start_producer/4

register_consumer(Client, Topic, Partition)

Specs

brod_client.register_consumer/3

Register self() as a partition consumer. The pid is registered in an ETS table, then the callers may lookup a consumer pid from the table ane make subscribe calls to the process directly.

register_producer(Client, Topic, Partition)

Specs

brod_client.register_producer/3

Register self() as a partition producer. The pid is registered in an ETS table, then the callers may lookup a producer pid from the table and make produce requests to the producer process directly.

request_sync(State, Request)

Specs

brod_client.request_sync/2

safe_gen_call(Server, Call, Timeout)

Specs

brod_client.safe_gen_call/3

shutdown_pid(Pid)

start_consumer(Client, TopicName, ConsumerConfig)

Specs

brod_client.start_consumer/3

Dynamically start a topic consumer. Returns ok if the consumer is already started.

start_link(BootstrapEndpoints, ClientId, Config)

Specs

brod_client.start_link/3

start_producer(Client, TopicName, ProducerConfig)

Specs

brod_client.start_producer/3

Dynamically start a per-topic producer. Return ok if the producer is already started.

stop(Client)

Specs

brod_client.stop/1

stop_consumer(Client, TopicName)

Specs

brod_client.stop_consumer/2

Stop all partition consumers of the given topic.

stop_producer(Client, TopicName)

Specs

brod_client.stop_producer/2

Stop all partition producers of the given topic.

timeout(State)

update_partitions_count_cache(Ets, Rest)

Specs

brod_client.update_partitions_count_cache/2

validate_topic_existence(Topic, State, IsRetry)

Specs

brod_client.validate_topic_existence/3

with_ok(Result, Continue)

Specs

brod_client.with_ok/2