brod_consumer (Brod v3.15.4)

Types

bytes()

Specs

t:brod_consumer.bytes/0

config()

Specs

t:brod_consumer.config/0

offset()

Specs

t:brod_consumer.offset/0

offset_reset_policy()

Specs

t:brod_consumer.offset_reset_policy/0

offset_time()

Specs

t:brod_consumer.offset_time/0

options()

Specs

t:brod_consumer.options/0

partition()

Specs

t:brod_consumer.partition/0

pending()

Specs

t:brod_consumer.pending/0

pending_acks()

Specs

t:brod_consumer.pending_acks/0

pending_queue()

Specs

t:brod_consumer.pending_queue/0

state()

Specs

t:brod_consumer.state/0

topic()

Specs

t:brod_consumer.topic/0

Functions

ack(Pid, Offset)

Specs

brod_consumer.ack/2

Subscriber confirms that a message (identified by offset) has been consumed, consumer process now may continue to fetch more messages.

add_pending_ack(Kafka_message, Pending_acks)

add_pending_acks(PendingAcks, Messages)

cast_to_subscriber(Pid, Msg)

connect_leader(ClientPid, Topic, Partition)

debug(Pid, File)

Specs

brod_consumer.debug/2

Enable/disable debugging on the consumer process.

debug(Pid, print) prints debug info to stdout.
debug(Pid, File) prints debug info to a file File.

do_debug(Pid, Debug)

err_op(_)

get_connection(Pid)

Get connection pid. Test/debug only.

handle_ack(Pending_acks, Offset)

handle_batches(Header, Batches, State)

handle_conn_down(State0)

handle_fetch_error(Kafka_fetch_error, State)

handle_fetch_response(Kpro_rsp, State)

handle_reset_offset(State, Error)

handle_subscribe_call(Pid, Options, State)

init(_)

Callback implementation for :gen_server.init/1.

is_shared_conn(Bootstrap)

maybe_delay_fetch_request(State)

Specs

brod_consumer.maybe_delay_fetch_request/1

maybe_init_connection(State)

Specs

brod_consumer.maybe_init_connection/1

maybe_send_fetch_request(State)

maybe_send_init_connection(State)

Specs

brod_consumer.maybe_send_init_connection/1

maybe_shrink_max_bytes(State, Messages)

reset_buffer(State)

Specs

brod_consumer.reset_buffer/1

resolve_begin_offset(State)

Specs

brod_consumer.resolve_begin_offset/1

resolve_offset(Connection, Topic, Partition, BeginOffset)

Specs

brod_consumer.resolve_offset/4

safe_gen_call(Server, Call, Timeout)

Specs

brod_consumer.safe_gen_call/3

send_fetch_request(State)

Specs

brod_consumer.send_fetch_request/1

start_link(Bootstrap, Topic, Partition, Config)

Specs

brod_consumer.start_link/4

Equivalent to start_link(ClientPid, Topic, Partition, Config, []).

start_link(Bootstrap, Topic, Partition, Config, Debug)

Specs

brod_consumer.start_link/5

Start (link) a partition consumer.
Possible configs:

min_bytes (optional, default = 0)
Minimal bytes to fetch in a batch of messages

max_bytes (optional, default = 1MB)

Maximum bytes to fetch in a batch of messages.
NOTE: this value might be expanded to retry when it is not enough to fetch even a single message, then slowly shrinked back to the given value.

max_wait_time (optional, default = 10000 ms)
Max number of seconds allowd for the broker to collect min_bytes of messages in fetch response

sleep_timeout (optional, default = 1000 ms)
Allow consumer process to sleep this amout of ms if kafka replied 'empty' message set.

prefetch_count (optional, default = 10)
The window size (number of messages) allowed to fetch-ahead.

prefetch_bytes (optional, default = 100KB)
The total number of bytes allowed to fetch-ahead. brod_consumer is greed, it only stops fetching more messages in when number of unacked messages has exceeded prefetch_count AND the unacked total volume has exceeded prefetch_bytes

begin_offset (optional, default = latest)
The offset from which to begin fetch requests.

offset_reset_policy (optional, default = reset_by_subscriber)

How to reset begin_offset if OffsetOutOfRange exception is received.

reset_by_subscriber: consumer is suspended (is_suspended=true in state) and wait for subscriber to re-subscribe with a new begin_offset option.

reset_to_earliest: consume from the earliest offset.
reset_to_latest: consume from the last available offset.

size_stat_window: (optional, default = 5)
The moving-average window size to caculate average message size. Average message size is used to shrink max_bytes in fetch requests after it has been expanded to fetch a large message. Use 0 to immediately shrink back to original max_bytes from config. A size estimation allows users to set a relatively small max_bytes, then let it dynamically adjust to a number around PrefetchCount * AverageSize

stop(Pid)

Specs

brod_consumer.stop/1

stop_maybe_kill(Pid, Timeout)

Specs

brod_consumer.stop_maybe_kill/2

subscribe(Pid, SubscriberPid, ConsumerOptions)

Specs

brod_consumer.subscribe/3

Subscribe or resubscribe on messages from a partition. Caller may specify a set of options extending consumer config. It is possible to update parameters such as max_bytes and max_wait_time, or the starting point (begin_offset) of the data stream.

Possible options:

All consumer configs as documented for start_link/5

begin_offset (optional, default = latest)
A subscriber may consume and process messages, then persist the associated offset to a persistent storage, then start (or restart) from last_processed_offset + 1 as the begin_offset to proceed. By default, it starts fetching from the latest available offset.

unsubscribe(Pid, SubscriberPid)

Specs

brod_consumer.unsubscribe/2

Unsubscribe the current subscriber.

update_avg_size(State, Rest)

update_options(Options, State)

Specs

brod_consumer.update_options/2