brod_group_coordinator (Brod v3.15.4)

Types

brod_offset_commit_policy()

Specs

t:brod_group_coordinator.brod_offset_commit_policy/0

brod_partition_assignment_strategy()

Specs

t:brod_group_coordinator.brod_partition_assignment_strategy/0

config()

Specs

t:brod_group_coordinator.config/0

member()

Specs

t:brod_group_coordinator.member/0

member_id()

Specs

t:brod_group_coordinator.member_id/0

offset_commit_policy()

Specs

t:brod_group_coordinator.offset_commit_policy/0

partition_assignment_strategy()

Specs

t:brod_group_coordinator.partition_assignment_strategy/0

protocol_name()

Specs

t:brod_group_coordinator.protocol_name/0

state()

Specs

t:brod_group_coordinator.state/0

ts()

Specs

t:brod_group_coordinator.ts/0

Functions

ack(Pid, GenerationId, Topic, Partition, Offset)

Specs

brod_group_coordinator.ack/5

For group member to call to acknowledge a consumed message offset.

all_topics(Members)

Specs

brod_group_coordinator.all_topics/1

assert_commit_response(Topics)

Specs

brod_group_coordinator.assert_commit_response/1

assign_partitions(State)

Specs

brod_group_coordinator.assign_partitions/1

bin(X)

Specs

brod_group_coordinator.bin/1

code_change(OldVsn, State, Extra)

Callback implementation for :gen_server.code_change/3.

collect_commit_response_error_codes(Topics)

Specs

brod_group_coordinator.collect_commit_response_error_codes/1

commit_offsets(CoordinatorPid)

Specs

brod_group_coordinator.commit_offsets/1

Force commit collected (acked) offsets immediately.

commit_offsets(CoordinatorPid, Offsets0)

Specs

brod_group_coordinator.commit_offsets/2

Force commit collected (acked) offsets plus the given extra offsets immediately.
NOTE: lists:usort/1 is applied on the given extra offsets to commit, meaning if two or more offsets for the same topic-partition exist in the list, only the one closest the head of the list is kept

coordinator_id()

Specs

brod_group_coordinator.coordinator_id/0

discover_coordinator(State)

Specs

brod_group_coordinator.discover_coordinator/1

do_assign_partitions(_, Members, AllPartitions)

Specs

brod_group_coordinator.do_assign_partitions/3

do_commit_offsets(State)

Specs

brod_group_coordinator.do_commit_offsets/1

do_commit_offsets_(State)

Specs

brod_group_coordinator.do_commit_offsets_/1

do_stabilize(Rest, RetryFun, State)

ensure_leader_at_hd(LeaderId, Members)

format_assignments(Assignments)

Specs

brod_group_coordinator.format_assignments/1

format_partition_assignments(Rest)

Specs

brod_group_coordinator.format_partition_assignments/1

get_committed_offsets(State, TopicPartitions)

Specs

brod_group_coordinator.get_committed_offsets/2

get_partitions(Client, Topic)

Specs

brod_group_coordinator.get_partitions/2

get_topic_assignments(State, Assignment)

Specs

brod_group_coordinator.get_topic_assignments/2

handle_ack(State, GenerationId, Topic, Partition, Offset)

Specs

brod_group_coordinator.handle_ack/5

handle_call(Call, From, State)

Callback implementation for :gen_server.handle_call/3.

handle_cast(Cast, State)

Callback implementation for :gen_server.handle_cast/2.

handle_info(Info, State)

Callback implementation for :gen_server.handle_info/2.

init(_)

Callback implementation for :gen_server.init/1.

is_already_connected(State, _)

is_default_offset_retention(_)

is_roundrobin_v1_commit(Metadata)

Specs

brod_group_coordinator.is_roundrobin_v1_commit/1

join_group(State)

Specs

brod_group_coordinator.join_group/1

log(State, Level, Fmt, Args)

make_group_connection_client_id()

Specs

brod_group_coordinator.make_group_connection_client_id/0

make_offset_commit_metadata()

Specs

brod_group_coordinator.make_offset_commit_metadata/0

maybe_reset_member_id(State, Reason)

maybe_send_heartbeat(State)

Specs

brod_group_coordinator.maybe_send_heartbeat/1

maybe_upgrade_from_roundrobin_v1(Offset, Metadata)

Specs

brod_group_coordinator.maybe_upgrade_from_roundrobin_v1/2

merge_acked_offsets(Offsets, Offsets)

Specs

brod_group_coordinator.merge_acked_offsets/2

receive_pending_acks(State)

Specs

brod_group_coordinator.receive_pending_acks/1

resolve_begin_offsets(Rest, CommittedOffsets, IsConsumerManaged)

Specs

brod_group_coordinator.resolve_begin_offsets/3

roundrobin_assign_loop(Partitions, PendingMembers, AssignedMembers)

send_sync(Connection, Request)

send_sync(Connection, Request, Timeout)

should_reset_member_id(_)

stabilize(State, AttemptNo, Reason)

Specs

brod_group_coordinator.stabilize/3

start_heartbeat_timer(HbRateSec)

Specs

brod_group_coordinator.start_heartbeat_timer/1

start_link(Client, GroupId, Topics, Config, CbModule, MemberPid)

Specs

brod_group_coordinator.start_link/6

Start a kafka consumer group coordinator.

Client: ClientId (or pid, but not recommended)

GroupId: Predefined globally unique (in a kafka cluster) binary string.

Topics: Predefined set of topic names to join the group.

CbModule: The module which implements group coordinator callbacks

MemberPid: The member process pid.

Config: The group coordinator configs in a proplist, possible entries:

partition_assignment_strategy (optional, default = roundrobin_v2)

Possible values:

roundrobin_v2 (topic-sticky)
Take all topic-offset (sorted topic_partition() list), assign one to each member in a roundrobin fashion. Only partitions in the subscription topic list are assiged.

callback_implemented
Call CbModule:assign_partitions/2 to assign partitions.

session_timeout_seconds (optional, default = 10)
Time in seconds for the group coordinator broker to consider a member 'down' if no heartbeat or any kind of requests received from a broker in the past N seconds. A group member may also consider the coordinator broker 'down' if no heartbeat response response received in the past N seconds.

heartbeat_rate_seconds (optional, default = 2)
Time in seconds for the member to 'ping' the group coordinator. OBS: Care should be taken when picking the number, on one hand, we do not want to flush the broker with requests if we set it too low, on the other hand, if set it too high, it may take too long for the members to realise status changes of the group such as assignment rebalacing or group coordinator switchover etc.

max_rejoin_attempts (optional, default = 5)
Maximum number of times allowd for a member to re-join the group. The gen_server will stop if it reached the maximum number of retries. OBS: 'let it crash' may not be the optimal strategy here because the group member id is kept in the gen_server looping state and it is reused when re-joining the group.

rejoin_delay_seconds (optional, default = 1)
Delay in seconds before re-joining the group.

offset_commit_policy (optional, default = commit_to_kafka_v2)
How/where to commit offsets, possible values:

commit_to_kafka_v2: Group coordinator will commit the offsets to kafka using version 2 OffsetCommitRequest.

consumer_managed: The group member (e.g. brod_group_subscriber.erl) is responsible for persisting offsets to a local or centralized storage. And the callback get_committed_offsets should be implemented to allow group coordinator to retrieve the commited offsets.

offset_commit_interval_seconds (optional, default = 5)
The time interval between two OffsetCommitRequest messages. This config is irrelevant if offset_commit_policy is consumer_managed.

offset_retention_seconds (optional, default = -1)
How long the time is to be kept in kafka before it is deleted. The default special value -1 indicates that the __consumer_offsets topic retention policy is used. This config is irrelevant if offset_commit_policy is consumer_managed.

protocol_name (optional, default = roundrobin_v2)
This is the protocol name used when join a group, if not given, by default partition_assignment_strategy is used as the protocol name. Setting a protocol name allows to interact with consumer group members designed in other programing languages. For example, 'range' is the most commonly used protocol name for JAVA client. However, brod only supports roundrobin protocol out of the box, in order to mimic 'range' protocol one will have to do it via callback_implemented assignment strategy

start_offset_commit_timer(State)

Specs

brod_group_coordinator.start_offset_commit_timer/1

stop(Pid)

Specs

brod_group_coordinator.stop/1

Stop group coordinator, wait for pid DOWN before return.

sync_group(State)

Specs

brod_group_coordinator.sync_group/1

terminate(Reason, State)

Callback implementation for :gen_server.terminate/2.

translate_members(Members)

Specs

brod_group_coordinator.translate_members/1

try_commit_offsets(State)

Specs

brod_group_coordinator.try_commit_offsets/1

update_topics(CoordinatorPid, Topics)

Specs

brod_group_coordinator.update_topics/2

Update the list of topics the brod_group_coordinator follow which triggers a join group rebalance

user_data(Module, Pid)