brod_group_subscriber behaviour (Brod v3.15.4)

A group subscriber is a gen_server which subscribes to partition consumers (poller) and calls the user-defined callback functions for message processing.
An overview of what it does behind the scene:

Start a consumer group coordinator to manage the consumer group states, see brod_group_coordinator:start_link/6

Start (if not already started) topic-consumers (pollers) and subscribe to the partition workers when group assignment is received from the group leader, see brod:start_consumer/3

Call CallbackModule:handle_message/4 when messages are received from the partition consumers.

Send acknowledged offsets to group coordinator which will be committed to kafka periodically.

Types

ack_ref()

Specs

t:brod_group_subscriber.ack_ref/0

cb_state()

Specs

t:brod_group_subscriber.cb_state/0

consumer()

Specs

t:brod_group_subscriber.consumer/0

member_id()

Specs

t:brod_group_subscriber.member_id/0

state()

Specs

t:brod_group_subscriber.state/0

Functions

ack(Pid, Topic, Partition, Offset)

Specs

brod_group_subscriber.ack/4

Acknowledge and commit an offset. The subscriber may ack a later (greater) offset which will be considered as multi-acking the earlier (smaller) offsets. This also means that disordered acks may overwrite offset commits and lead to unnecessary message re-delivery in case of restart.

ack(Pid, Topic, Partition, Offset, Commit)

Specs

brod_group_subscriber.ack/5

Acknowledge an offset. This call may or may not commit group subscriber offset depending on the value of Commit argument

assign_partitions(Pid, Members, TopicPartitionList)

Specs

brod_group_subscriber.assign_partitions/3

This function is called only when partition_assignment_strategy is set for callback_implemented in group config.

assignments_received(Pid, MemberId, GenerationId, TopicAssignments)

Specs

brod_group_subscriber.assignments_received/4

Called by group coordinator when there is new assignment received.

assignments_revoked(Pid)

Specs

brod_group_subscriber.assignments_revoked/1

Called by group coordinator before re-joining the consumer group.

code_change(OldVsn, State, Extra)

Callback implementation for :gen_server.code_change/3.

commit(Pid)

Specs

brod_group_subscriber.commit/1

Commit all acked offsets. NOTE: This is an async call.

commit(Pid, Topic, Partition, Offset)

Specs

brod_group_subscriber.commit/4

Commit offset for a topic. This is an asynchronous call

consume_ack(Pid, Offset)

do_commit_ack(Pid, GenerationId, Topic, Partition, Offset)

get_committed_offsets(Pid, TopicPartitions)

Specs

brod_group_subscriber.get_committed_offsets/2

Called by group coordinator when initializing the assignments for subscriber.

NOTE: This function is called only when offset_commit_policy is set to consumer_managed in group config.
NOTE: The committed offsets should be the offsets for successfully processed (acknowledged) messages, not the begin_offset to start fetching from.

get_consumer(Pid, Consumers)

handle_ack(AckRef, State, CommitNow)

Specs

brod_group_subscriber.handle_ack/3

handle_call(Call, From, State)

Callback implementation for :gen_server.handle_call/3.

handle_cast(Cast, State)

Callback implementation for :gen_server.handle_cast/2.

handle_consumer_delivery(Kafka_message_set, State)

handle_info(Info, State0)

Callback implementation for :gen_server.handle_info/2.

handle_message_set(MessageSet, State)

handle_messages(Topic, Partition, Rest, State)

init(_)

Callback implementation for :gen_server.init/1.

log(State, Level, Fmt, Args)

put_consumer(Consumer, Consumers)

start_link(Client, GroupId, Topics, GroupConfig, ConsumerConfig, CbModule, CbInitArg)

Specs

brod_group_subscriber.start_link/7

Equivalent to start_link(Client, GroupId, Topics, GroupConfig, ConsumerConfig, message, CbModule, CbInitArg).

start_link(Client, GroupId, Topics, GroupConfig, ConsumerConfig, MessageType, CbModule, CbInitArg)

Specs

brod_group_subscriber.start_link/8

Start (link) a group subscriber.

Client: Client ID (or pid, but not recommended) of the brod client.

GroupId: Consumer group ID which should be unique per kafka cluster

Topics: Predefined set of topic names to join the group.

NOTE: The group leader member will collect topics from all members and assign all collected topic-partitions to members in the group. i.e. members can join with arbitrary set of topics.

GroupConfig: For group coordinator, see brod_group_coordinator:start_link/6

ConsumerConfig: For partition consumer, see brod_consumer:start_link/4

MessageType: The type of message that is going to be handled by the callback module. Can be either message or message_set.

CbModule: Callback module which should have the callback functions implemented for message processing.
CbInitArg: The term() that is going to be passed to CbModule:init/2 as a second argument when initializing the subscriber.

start_subscribe_timer(Ref, Delay)

Specs

brod_group_subscriber.start_subscribe_timer/2

stop(Pid)

Specs

brod_group_subscriber.stop/1

Stop group subscriber, wait for pid DOWN before return.

subscribe_partition(Client, Consumer)

subscribe_partitions(State)

terminate(Reason, State)

Callback implementation for :gen_server.terminate/2.

update_last_offset(Messages, Consumer0, Consumers)

user_data(Pid)

Callback implementation for c::brod_group_member.user_data/1.

Callbacks

handle_message(arg1, arg2, arg3, cb_state)

Specs

c:brod_group_subscriber.handle_message/4

init(arg1, term)

Specs

c:brod_group_subscriber.init/2