brod_group_subscriber_v2 behaviour (Brod v3.15.4)

This module implements an improved version of brod_group_subscriber behavior. Key difference is that each partition worker runs in a separate Erlang process, allowing parallel message processing.

Types

commit_fun()

Specs

t:brod_group_subscriber_v2.commit_fun/0

committed_offsets()

Specs

t:brod_group_subscriber_v2.committed_offsets/0

init_info()

Specs

t:brod_group_subscriber_v2.init_info/0

member_id()

Specs

t:brod_group_subscriber_v2.member_id/0

state()

Specs

t:brod_group_subscriber_v2.state/0

subscriber_config()

Specs

t:brod_group_subscriber_v2.subscriber_config/0

worker()

Specs

t:brod_group_subscriber_v2.worker/0

workers()

Specs

t:brod_group_subscriber_v2.workers/0

Functions

ack(Pid, Topic, Partition, Offset)

Specs

brod_group_subscriber_v2.ack/4

Commit offset for a topic-partition, but don't commit it to Kafka. This is an asynchronous call

assign_partitions(Pid, Members, TopicPartitionList)

Specs

brod_group_subscriber_v2.assign_partitions/3

This function is called only when partition_assignment_strategy is set for callback_implemented in group config.

assignments_received(Pid, MemberId, GenerationId, TopicAssignments)

Specs

brod_group_subscriber_v2.assignments_received/4

Called by group coordinator when there is new assignment received.

assignments_revoked(Pid)

Specs

brod_group_subscriber_v2.assignments_revoked/1

Called by group coordinator before re-joining the consumer group.

commit(Pid, Topic, Partition, Offset)

Specs

brod_group_subscriber_v2.commit/4

Ack offset for a topic-partition. This is an asynchronous call

do_ack(Topic, Partition, Offset, State)

Specs

brod_group_subscriber_v2.do_ack/4

flush_offset_commits(GroupId, Coordinator)

get_committed_offsets(Pid, TopicPartitions)

Specs

brod_group_subscriber_v2.get_committed_offsets/2

Called by group coordinator when initializing the assignments for subscriber.

NOTE: This function is called only when offset_commit_policy is set to consumer_managed in group config.
NOTE: The committed offsets should be the offsets for successfully processed (acknowledged) messages, not the begin_offset to start fetching from.

handle_call(Call, From, State)

Callback implementation for :gen_server.handle_call/3.

handle_worker_failure(_, Pid, Reason, State)

Specs

brod_group_subscriber_v2.handle_worker_failure/4

maybe_start_worker(MemberId, ConsumerConfig, Topic, Partition, BeginOffset, State)

Specs

brod_group_subscriber_v2.maybe_start_worker/6

start_link(Config)

Specs

brod_group_subscriber_v2.start_link/1

Start (link) a group subscriber.

Possible Config keys:

client: Client ID (or pid, but not recommended) of the brod client. Mandatory

group_id: Consumer group ID which should be unique per kafka cluster. Mandatory

topics: Predefined set of topic names to join the group. Mandatory
NOTE: The group leader member will collect topics from all members and assign all collected topic-partitions to members in the group. i.e. members can join with arbitrary set of topics.

cb_module: Callback module which should have the callback functions implemented for message processing. Mandatory

group_config: For group coordinator, see brod_group_coordinator:start_link/6 Optional

consumer_config: For partition consumer, brod_topic_subscriber:start_link/6. Optional

message_type: The type of message that is going to be handled by the callback module. Can be either message or message set. Optional, defaults to message

init_data: The term() that is going to be passed to CbModule:init/2 when initializing the subscriber. Optional, defaults to undefined

start_worker(Client, Topic, MessageType, Partition, ConsumerConfig, StartOptions)

Specs

brod_group_subscriber_v2.start_worker/6

stop(Pid)

Specs

brod_group_subscriber_v2.stop/1

Stop group subscriber, wait for pid DOWN before return.

terminate_all_workers(Workers)

Specs

brod_group_subscriber_v2.terminate_all_workers/1

terminate_worker(WorkerPid)

Specs

brod_group_subscriber_v2.terminate_worker/1

Callbacks

assign_partitions(_cbConfig, list, list)
(optional)

Specs

c:brod_group_subscriber_v2.assign_partitions/3

get_committed_offset(_cbConfig, arg2, arg3)
(optional)

Specs

c:brod_group_subscriber_v2.get_committed_offset/3

handle_message(arg1, state)

Specs

c:brod_group_subscriber_v2.handle_message/2

init(arg1, _cbConfig)

Specs

c:brod_group_subscriber_v2.init/2

terminate(_reason, _state)
(optional)

Specs

c:brod_group_subscriber_v2.terminate/2