brod_kafka_request (Brod v3.15.4)

Helper functions for building request messages.

Types

api()

Specs

t:brod_kafka_request.api/0

conn()

Specs

t:brod_kafka_request.conn/0

offset()

Specs

t:brod_kafka_request.offset/0

partition()

Specs

t:brod_kafka_request.partition/0

topic()

Specs

t:brod_kafka_request.topic/0

topic_config()

Specs

t:brod_kafka_request.topic_config/0

vsn()

Specs

t:brod_kafka_request.vsn/0

Functions

create_topics(Connection, TopicConfigs, RequestConfigs)

Specs

brod_kafka_request.create_topics/3

Make a create_topics request.

delete_topics(Connection, Topics, Timeout)

Specs

brod_kafka_request.delete_topics/3

Make a delete_topics request.

ensure_integer_offset_time(T)

Specs

brod_kafka_request.ensure_integer_offset_time/1

fetch(Pid, Topic, Partition, Offset, WaitTime, MinBytes, MaxBytes)

Specs

brod_kafka_request.fetch/7

Make a fetch request, If the first arg is a connection pid, call brod_kafka_apis:pick_version/2 to resolve version.

join_group(Conn, Fields)

Specs

brod_kafka_request.join_group/2

Make a join_group request.

list_groups(Connection)

Specs

brod_kafka_request.list_groups/1

Make a list_groups request.

list_offsets(Connection, Topic, Partition, TimeOrSemanticOffset)

Specs

brod_kafka_request.list_offsets/4

Make a list_offsets request message for offset resolution. In kafka protocol, -2 and -1 are semantic 'time' to request for 'earliest' and 'latest' offsets. In brod implementation, -2, -1, 'earliest' and 'latest' are semantic 'offset', this is why often a variable named Offset is used as the Time argument.

make_req(API, Conn, Fields)

metadata(Connection, Topics)

Specs

brod_kafka_request.metadata/2

Make a metadata request.

offset_commit(Conn, Fields)

Specs

brod_kafka_request.offset_commit/2

Make a offset_commit request.

offset_fetch(Connection, GroupId, Topics)

Specs

brod_kafka_request.offset_fetch/3

Make a offset fetch request. NOTE: empty topics list only works for kafka 0.10.2.0 or later

pick_version(API, Vsn)

Specs

brod_kafka_request.pick_version/2

produce(MaybePid, Topic, Partition, BatchInput, RequiredAcks, AckTimeout, Compression)

Specs

brod_kafka_request.produce/7

Make a produce request, If the first arg is a connection pid, call brod_kafka_apis:pick_version/2 to resolve version.

sync_group(Conn, Fields)

Specs

brod_kafka_request.sync_group/2

Make a sync_group request.