brod_topic_subscriber behaviour (Brod v3.15.4)

A topic subscriber is a gen_server which subscribes to all or a given set of partition consumers (pollers) of a given topic and calls the user-defined callback functions for message processing.

Types

ack_ref()

Specs

t:brod_topic_subscriber.ack_ref/0

cb_fun()

Specs

t:brod_topic_subscriber.cb_fun/0

cb_ret()

Specs

t:brod_topic_subscriber.cb_ret/0

cb_state()

Specs

t:brod_topic_subscriber.cb_state/0

committed_offsets()

Specs

t:brod_topic_subscriber.committed_offsets/0

consumer()

Specs

t:brod_topic_subscriber.consumer/0

state()

Specs

t:brod_topic_subscriber.state/0

topic_subscriber_config()

Specs

t:brod_topic_subscriber.topic_subscriber_config/0

Functions

ack(Pid, Partition, Offset)

Specs

brod_topic_subscriber.ack/3

Acknowledge that message has been sucessfully consumed.

consume_ack(Pid, Offset)

get_consumer(Partition, Consumers)

handle_ack(AckRef, State)

Specs

brod_topic_subscriber.handle_ack/2

handle_consumer_delivery(Kafka_message_set, State)

handle_message_set(MessageSet, State)

handle_messages(Partition, Rest, State)

put_consumer(Consumer, Consumers)

send_lo_cmd(CMD)

send_lo_cmd(CMD, DelayMS)

start_link(Config)

Specs

brod_topic_subscriber.start_link/1

Start (link) a topic subscriber which receives and processes the messages from a given partition set.

Possible Config keys:

client: Client ID (or pid, but not recommended) of the brod client. Mandatory

topic: Topic to consume from. Mandatory

cb_module: Callback module which should have the callback functions implemented for message processing. Mandatory

consumer_config: For partition consumer, brod_topic_subscriber:start_link/6. Optional, defaults to []

message_type: The type of message that is going to be handled by the callback module. Can be either message or message set. Optional, defaults to message_set

init_data: The term() that is going to be passed to CbModule:init/2 when initializing the subscriber. Optional, defaults to undefined

partitions: List of partitions to consume from, or atom all. Optional, defaults to all.

start_link(Client, Topic, Partitions, ConsumerConfig, CbModule, CbInitArg)

This function is deprecated. Please use start_link/1 instead.

Specs

brod_topic_subscriber.start_link/6

Equivalent to start_link(Client, Topic, Partitions, ConsumerConfig, message, CbModule, CbInitArg).

start_link(Client, Topic, Partitions, ConsumerConfig, MessageType, CbModule, CbInitArg)

This function is deprecated. Please use start_link/1 instead.

Specs

brod_topic_subscriber.start_link/7

Start (link) a topic subscriber which receives and processes the messages or message sets from the given partition set. Use atom all to subscribe to all partitions. Messages are handled by calling CbModule:handle_message

start_link(Client, Topic, Partitions, ConsumerConfig, CommittedOffsets, MessageType, CbFun, CbInitialState)

This function is deprecated. Please use start_link/1 instead.

Specs

brod_topic_subscriber.start_link/8

Start (link) a topic subscriber which receives and processes the messages from the given partition set. Use atom all to subscribe to all partitions. Messages are handled by calling the callback function.
NOTE: CommittedOffsets are the offsets for the messages that have been successfully processed (acknowledged), not the begin-offset to start fetching from.

stop(Pid)

Specs

brod_topic_subscriber.stop/1

Stop topic subscriber.

subscribe_partition(Client, Topic, Consumer)

subscribe_partitions(State)

update_last_offset(Partition, Messages, State)

Callbacks

handle_message(arg1, arg2, cb_state)

Specs

c:brod_topic_subscriber.handle_message/3

init(arg1, term)

Specs

c:brod_topic_subscriber.init/2

terminate(_reason, cb_state)
(optional)

Specs

c:brod_topic_subscriber.terminate/2